gfx_camera.cpp: use integer offset for small-screen cam

+++ src/graphics/gfx_camera.cpp
@@ -104,7 +104,7 @@ void ProcessSmallScreenCam(vScreen_t& vscreen)
                 vscreen.small_screen_features.offset_x -= rateX;
         }
 
-        vscreen.X -= rate * vscreen.small_screen_features.offset_x / 2;
+        vscreen.X -= (int)(rate * vscreen.small_screen_features.offset_x / 2);
     }
 
     if(g_config.small_screen_cam && screen.H < c_screen.H)
@@ -192,7 +192,7 @@ void ProcessSmallScreenCam(vScreen_t& vscreen)
         fix_default_offset *= (c_screen.H - screen.H) - std::abs(lookY) * 2;
         fix_default_offset /= (c_screen.H - screen.H);
 
-        vscreen.Y += rate * (lookY + fix_default_offset);
+        vscreen.Y += (int)(rate * (lookY + fix_default_offset));
     }
 }
